requirements.txt
Django==2.2.8
psycopg2-binary==2.8.4
python-escpos==2.2.0
Jinja2==2.10.3
markupsafe==2.0.1 # for jinja2
django-bootstrap-modal-forms==1.5.0
django-pure-pagination==0.3.0
xlwt==1.3.0
django-xhtml2pdf==0.0.4
django-widget-tweaks==1.4.8
django-crispy-forms==1.9.1
django-simple-history==2.8.0
django-model-utils==4.0.0
djangorestframework==3.11.0
django-mptt==0.11.0
django-extensions==2.2.9
xlrd==1.2.0
lxml==4.5.0
xmlsig==0.1.3
pyOpenSSL==19.1.0
suds-jurko==0.6
requests==2.23.0
django-mathfilters==1.0.0
